+++ src/generator.rs
@@ -24,7 +24,7 @@ pub use default_rng::DefaultRng;
 /// # Examples
 ///
 /// ```rust
-/// # #[cfg(feature = "std")]
+/// # #[cfg(feature = "default_rng")]
 /// # {
 /// use scru128::Scru128Generator;
 ///
@@ -39,7 +39,7 @@ pub use default_rng::DefaultRng;
 /// synchronization mechanisms to control the scope of guaranteed monotonicity:
 ///
 /// ```rust
-/// # #[cfg(feature = "std")]
+/// # #[cfg(feature = "default_rng")]
 /// # {
 /// use scru128::Scru128Generator;
 /// use std::sync::{Arc, Mutex};
@@ -106,7 +106,7 @@ impl<R: Scru128Rng> Scru128Generator<R> {
     /// # Examples
     ///
     /// ```rust
-    /// # #[cfg(feature = "std")]
+    /// # #[cfg(feature = "default_rng")]
     /// # {
     /// use scru128::Scru128Generator;
     ///
@@ -204,8 +204,8 @@ impl<R: Scru128Rng> Scru128Generator<R> {
     }
 }
 
-#[cfg(any(feature = "std", test))]
-#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
+#[cfg(any(feature = "default_rng", test))]
+#[cfg_attr(docsrs, doc(cfg(feature = "default_rng")))]
 impl Scru128Generator {
     /// Creates a generator object with the default random number generator.
     pub fn new() -> Self {
@@ -247,6 +247,8 @@ mod std_ext {
         /// # Examples
         ///
         /// ```rust
+        /// # #[cfg(feature = "default_rng")]
+        /// # {
         /// use scru128::Scru128Generator;
         ///
         /// let mut g = Scru128Generator::new();
@@ -255,6 +257,7 @@ mod std_ext {
         ///     .generate_or_abort()
         ///     .expect("The clock went backwards by ten seconds!");
         /// assert!(x < y);
+        /// # }
         /// ```
         pub fn generate_or_abort(&mut self) -> Option<Scru128Id> {
             self.generate_or_abort_core(unix_ts_ms(), DEFAULT_ROLLBACK_ALLOWANCE)
@@ -267,12 +270,15 @@ mod std_ext {
     /// # Examples
     ///
     /// ```rust
+    /// # #[cfg(feature = "default_rng")]
+    /// # {
     /// use scru128::Scru128Generator;
     ///
     /// let g = Scru128Generator::new();
     /// for (i, e) in g.take(8).enumerate() {
     ///     println!("[{i}] {e}");
     /// }
+    /// # }
     /// ```
     impl<R: Scru128Rng> Iterator for Scru128Generator<R> {
         type Item = Scru128Id;
@@ -387,10 +393,10 @@ mod tests_generate_or_abort {
 }
 
 mod default_rng {
-    #[cfg(feature = "std")]
+    #[cfg(feature = "default_rng")]
     use rand::{rngs::adapter::ReseedingRng, rngs::OsRng, SeedableRng as _};
 
-    #[cfg(all(not(feature = "std"), test))]
+    #[cfg(all(not(feature = "default_rng"), test))]
     use rand::{rngs::StdRng, SeedableRng as _};
 
     /// The default random number generator used by [`Scru128Generator`].
@@ -400,47 +406,47 @@ mod default_rng {
     /// the same strategy as that employed by [`ThreadRng`]; see the docs of `rand` crate for a
     /// detailed discussion on the strategy.
     ///
-    /// This structure does exist under the `no_std` environment but is not able to be instantiated
-    /// or used as a random number generator.
+    /// This structure does exist without the `default_rng` feature flag but is not able to be
+    /// instantiated or used as a random number generator.
     ///
     /// [`Scru128Generator`]: super::Scru128Generator
     /// [`ChaCha12Core`]: rand_chacha::ChaCha12Core
     /// [`OsRng`]: rand::rngs::OsRng
     /// [`ReseedingRng`]: rand::rngs::adapter::ReseedingRng
     /// [`ThreadRng`]: https://docs.rs/rand/0.8/rand/rngs/struct.ThreadRng.html
-    #[cfg_attr(docsrs, doc(cfg(feature = "std")))]
+    #[cfg_attr(docsrs, doc(cfg(feature = "default_rng")))]
     #[derive(Clone, Debug)]
     pub struct DefaultRng {
         _private: (),
 
-        #[cfg(feature = "std")]
+        #[cfg(feature = "default_rng")]
         inner: ReseedingRng<rand_chacha::ChaCha12Core, OsRng>,
 
-        #[cfg(all(not(feature = "std"), test))]
+        #[cfg(all(not(feature = "default_rng"), test))]
         inner: StdRng,
     }
 
-    #[cfg(any(feature = "std", test))]
+    #[cfg(any(feature = "default_rng", test))]
     impl super::Scru128Rng for DefaultRng {
         fn next_u32(&mut self) -> u32 {
             rand::RngCore::next_u32(&mut self.inner)
         }
     }
 
-    #[cfg(any(feature = "std", test))]
+    #[cfg(any(feature = "default_rng", test))]
     impl Default for DefaultRng {
         fn default() -> Self {
             Self {
                 _private: (),
 
-                #[cfg(feature = "std")]
+                #[cfg(feature = "default_rng")]
                 inner: {
                     let rng = rand_chacha::ChaCha12Core::from_rng(OsRng)
                         .expect("could not initialize DefaultRng");
                     ReseedingRng::new(rng, 1024 * 64, OsRng)
                 },
 
-                #[cfg(all(not(feature = "std"), test))]
+                #[cfg(all(not(feature = "default_rng"), test))]
                 inner: {
                     let local_var = 0u32;
                     let addr_as_seed = (&local_var as *const u32) as u64;