+++ CHANGELOG.md
@@ -28,6 +28,7 @@ a case-insensitive scheme. Other changes include the removal of deprecated APIs.
 ### Added
 
 - `rand` feature flag to opt out dependency on `rand` crate
+- `default_rng` feature flag to opt out default `DefaultRng`
 - `global_gen` feature flag to opt out default global generator
 - `const` qualifier to `Scru128Id::encode()`
 - `Scru128Id::as_bytes()` and `impl AsRef<[u8]>` for `Scru128Id`